rename protocol extension module
